x86_64: regs: Activate Long Mode
When LME and CR0.PG are set, both LMA and CR4.PAE should be set as well.
As a matter of fact, recent kernels are checking for that condition when
setting KVM segment registers.

Without this patch crosvm won't boot guests on top of Linux kernels that
are more recent than v4.15-rc5.

TEST=crosvm boots on Linux v4.15.4
BUG=825221

Signed-off-by: Samuel Ortiz <samuel.ortiz@intel.com>